+++ src/main/java/chess/domain/piece/PieceRelation.java
@@ -1,6 +1,4 @@
-package chess.domain.position;
-
-import chess.domain.piece.Piece;
+package chess.domain.piece;
 
 public enum PieceRelation {
     NONE,
