+++ src/main/java/chess/domain/piece/type/Pawn.java
@@ -4,7 +4,7 @@ import chess.domain.piece.Piece;
 import chess.domain.piece.PieceColor;
 import chess.domain.position.ChessDirection;
 import chess.domain.position.ChessRank;
-import chess.domain.position.PieceRelation;
+import chess.domain.piece.PieceRelation;
 import chess.domain.position.Position;
 
 import java.util.Set;
