+++ src/main/java/chess/dto/CommandInfo.java
@@ -1,5 +1,7 @@
 package chess.dto;
 
+import chess.view.CommandType;
+
 import java.util.List;
 
 public record CommandInfo(CommandType type, List<String> arguments) {
